+++ conclave-common/src/main/kotlin/com/r3/conclave/common/internal/CallInterfaceMessageType.kt
@@ -2,20 +2,19 @@ package com.r3.conclave.common.internal
 
 /**
  * This forms part of the messaging protocol for enclave and host calls in native mode.
- * Messages (ecalls & ocalls) may represent calls, returns or exceptions.
+ * Messages (ECalls & OCalls) may represent calls, returns or exceptions.
  */
 enum class CallInterfaceMessageType {
     CALL,
     RETURN,
     EXCEPTION;
 
-    fun toByte(): Byte {
-        return ordinal.toByte()
-    }
+    fun toByte(): Byte = ordinal.toByte()
 
     companion object {
         private val VALUES = CallInterfaceMessageType.values()
 
+        @JvmStatic
         fun fromByte(byte: Byte): CallInterfaceMessageType {
             return VALUES[byte.toInt()]
         }