+++ conclave-common/src/main/kotlin/com/r3/conclave/common/internal/HostCallType.kt
@@ -10,14 +10,12 @@ enum class HostCallType {
     SET_ENCLAVE_INFO,
     CALL_MESSAGE_HANDLER;
 
-    fun toByte(): Byte {
-        check(VALUES.size < Byte.MAX_VALUE)
-        return ordinal.toByte()
-    }
+    fun toByte(): Byte = ordinal.toByte()
 
     companion object {
         private val VALUES = HostCallType.values()
 
+        @JvmStatic
         fun fromByte(i: Byte): HostCallType {
             require(i < VALUES.size) { "$i does not correspond to a valid host call type." }
             return VALUES[i.toInt()]