+++ conclave-enclave/src/main/kotlin/com/r3/conclave/enclave/internal/MockEnclaveEnvironment.kt
@@ -1,7 +1,7 @@
 package com.r3.conclave.enclave.internal
 
 import com.r3.conclave.common.EnclaveMode
-import com.r3.conclave.common.MockCallInterfaceConnector
+import com.r3.conclave.common.internal.MockCallInterfaceConnector
 import com.r3.conclave.common.MockConfiguration
 import com.r3.conclave.common.internal.*
 import com.r3.conclave.common.internal.KeyName.REPORT
@@ -26,8 +26,8 @@ class MockEnclaveEnvironment(
 ) : EnclaveEnvironment(loadEnclaveProperties(enclave::class.java, true), kdsConfig) {
     companion object {
         private fun versionToCpuSvn(num: Int): ByteArray {
-            return digest("SHA-256") { 
-                update(ByteBuffer.allocate(2).putShort(num.toShort()).array()) 
+            return digest("SHA-256") {
+                update(ByteBuffer.allocate(2).putShort(num.toShort()).array())
             }.copyOf(SgxCpuSvn.size)
         }