+++ conclave-host/src/main/kotlin/com/r3/conclave/host/internal/HostEnclaveInterface.kt
@@ -14,7 +14,7 @@ abstract class HostEnclaveInterface : CallInterface<EnclaveCallType, HostCallTyp
      * This is not currently used in mock mode.
      */
     fun initializeEnclave(enclaveClassName: String) {
-        initiateOutgoingCall(EnclaveCallType.INITIALISE_ENCLAVE, ByteBuffer.wrap(enclaveClassName.toByteArray(StandardCharsets.UTF_8)))
+        executeOutgoingCall(EnclaveCallType.INITIALISE_ENCLAVE, ByteBuffer.wrap(enclaveClassName.toByteArray()))
     }
 
     /**
@@ -26,21 +26,21 @@ abstract class HostEnclaveInterface : CallInterface<EnclaveCallType, HostCallTyp
             putNullable(sealedState) { put(it) }
             rewind()
         }
-        initiateOutgoingCall(EnclaveCallType.START_ENCLAVE, sealedStateBuffer)
+        executeOutgoingCall(EnclaveCallType.START_ENCLAVE, sealedStateBuffer)
     }
 
     /**
      * Stops the enclave, calling the onShutdown hook.
      */
     fun stopEnclave() {
-        initiateOutgoingCall(EnclaveCallType.STOP_ENCLAVE)
+        executeOutgoingCall(EnclaveCallType.STOP_ENCLAVE)
     }
 
     /**
      * Request a quote for enclave instance info from the enclave.
      */
     fun getEnclaveInstanceInfoQuote(target: ByteCursor<SgxTargetInfo>): ByteCursor<SgxSignedQuote> {
-        val returnBuffer = initiateOutgoingCallAndCheckReturn(EnclaveCallType.GET_ENCLAVE_INSTANCE_INFO_QUOTE, target.buffer)
+        val returnBuffer = executeOutgoingCallWithReturn(EnclaveCallType.GET_ENCLAVE_INSTANCE_INFO_QUOTE, target.buffer)
         return Cursor.wrap(SgxSignedQuote, returnBuffer.getRemainingBytes())
     }
 
@@ -49,7 +49,7 @@ abstract class HostEnclaveInterface : CallInterface<EnclaveCallType, HostCallTyp
      * Returns null if no KDS key spec is present in the enclave.
      */
     fun getKdsPersistenceKeySpec(): KDSKeySpec? {
-        return initiateOutgoingCall(EnclaveCallType.GET_KDS_PERSISTENCE_KEY_SPEC)?.let { buffer ->
+        return executeOutgoingCall(EnclaveCallType.GET_KDS_PERSISTENCE_KEY_SPEC)?.let { buffer ->
             val name = buffer.getIntLengthPrefixString()
             val masterKeyType = MasterKeyType.fromID(buffer.get().toInt())
             val policyConstraint = buffer.getRemainingString()
@@ -64,13 +64,13 @@ abstract class HostEnclaveInterface : CallInterface<EnclaveCallType, HostCallTyp
         val kdsResponseBuffer = ByteBuffer.allocate(kdsResponse.size).apply {
             putKdsPrivateKeyResponse(kdsResponse)
         }
-        initiateOutgoingCall(EnclaveCallType.SET_KDS_PERSISTENCE_KEY, kdsResponseBuffer)
+        executeOutgoingCall(EnclaveCallType.SET_KDS_PERSISTENCE_KEY, kdsResponseBuffer)
     }
 
     /**
      * Send a command to the enclave message handler.
      */
     fun sendMessageHandlerCommand(command: ByteBuffer) {
-        initiateOutgoingCall(EnclaveCallType.CALL_MESSAGE_HANDLER, command)
+        executeOutgoingCall(EnclaveCallType.CALL_MESSAGE_HANDLER, command)
     }
 }